t234: nv-public: introduce bwmgr device node
This device node represents a virtual device in the kernel
that would be used for exposing the control interface of EMC
clock floor frequency via linux devfreq framework.

EMC clock handle is needed to construct the OPP table, which
is used for devfreq framework to save the valid frequency
steps.

BPMP handle is needed to allow the driver to communicate with
the BPMP via mailbox infrastructure.

Bug 5555075

Change-Id: Ifdf1aedbfbeb1f875454370d8b00854a184b331d
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t23x-public-dts/+/3463467
Tested-by: Johnny Liu <johnliu@nvidia.com>
Reviewed-by: Johnny Liu <johnliu@nvidia.com>
Reviewed-by: Rajkumar Kasirajan <rkasirajan@nvidia.com>
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>